% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Bal_XeeXhC}
\alias{Bal_XeeXhC}
\title{Calculate Correlation Matrix for Balanced Panel}
\usage{
Bal_XeeXhC(dmat, X, e, n1, k)
}
\arguments{
\item{dmat}{distance matrix}

\item{X}{model matrix}

\item{e}{vector of residuals}

\item{n1}{number of observations}

\item{k}{number of regressors}
}
\description{
Calculate Correlation Matrix for Balanced Panel
}
