% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_helpers.R
\name{covgm_range}
\alias{covgm_range}
\title{Title}
\usage{
covgm_range(
  df.input,
  depvar = "noise1",
  indepvar = "noise2",
  maxdist = NA,
  spacing = NA,
  single.variable = FALSE
)
}
\arguments{
\item{single.variable}{}
}
\description{
Title
}
