% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DistMat}
\alias{DistMat}
\title{Create Distance Matrix}
\usage{
DistMat(M, cutoff, kernel = "bartlett", dist_fn = "Haversine")
}
\arguments{
\item{M}{a matrix of locations}

\item{cutoff}{the distance for cutoff}

\item{kernel}{(string) kernel function (default is bartlett-triangular)}

\item{dist_fn}{(string) distance function (Haversine)}
}
\description{
Create Distance Matrix
}
