% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_corr_range.R
\name{extract.corr.range}
\alias{extract.corr.range}
\title{Extract Correlation Range}
\usage{
extract.corr.range(ncf_corr)
}
\arguments{
\item{ncf_corr}{a correlogram estimated by ncf::correlog()}
}
\value{
an integer that represents the correlation range (in units of input, i.e. with lon lat it is km, with equal area CRS it tipically is metres)
}
\description{
Function will pick the point where the correlogram first crosses the x-axis (the distance), i.e. where the spatial autocorrelation gets 0 for the first time.
This is a good approximation of the actual correlation range.
}
