% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_sac.R
\docType{data}
\name{gm.param}
\alias{gm.param}
\title{Goodness-of-Fit Mapping for modelsummary}
\format{
An object of class \code{list} of length 5.
}
\usage{
gm.param
}
\description{
A list specifying how goodness-of-fit statistics should be displayed
in \code{\link[modelsummary:modelsummary]{modelsummary::modelsummary()}} output.
}
\keyword{internal}
