% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_as_cols.R
\name{sfc_as_cols}
\alias{sfc_as_cols}
\title{Coordinates as columns in dataframe}
\usage{
sfc_as_cols(x, names = c("x", "y"))
}
\arguments{
\item{names}{}
}
\value{

}
\description{
code comes from the internet
}
\examples{


}
