% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialInference-package.R
\docType{package}
\name{SpatialInference-package}
\alias{SpatialInference}
\alias{SpatialInference-package}
\title{SpatialInference: Conley Spatial HAC Standard Errors and Correlation Range Estimation}
\description{
Fast computation of Conley (1999) \doi{10.1016/S0304-4076(98)00084-0} spatial heteroskedasticity and autocorrelation consistent (HAC) standard errors for linear regression models with geo-coded data. Performance-critical distance calculations, kernel weighting, and variance component accumulation are implemented in C++ via 'Rcpp' and 'RcppArmadillo'. Includes tools for estimating the spatial correlation range from covariograms and correlograms, and diagnostic visualizations for bandwidth selection.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/axlehner/SpatialInference}
  \item Report bugs at \url{https://github.com/axlehner/SpatialInference/issues}
}

}
\author{
\strong{Maintainer}: Alexander Lehner \email{lehner@uchicago.edu} (\href{https://orcid.org/0000-0001-5885-5966}{ORCID})

}
\keyword{internal}
