% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TimeDist}
\alias{TimeDist}
\title{Temporal Variance Component (Time Dimension)}
\usage{
TimeDist(times, cutoff, X, e, n1, k)
}
\arguments{
\item{times}{numeric vector of time period identifiers}

\item{cutoff}{the lag cutoff (number of time periods)}

\item{X}{model matrix (n x k)}

\item{e}{vector of residuals (length n)}

\item{n1}{number of observations}

\item{k}{number of regressors}
}
\value{
A k x k matrix representing the temporal X'ee'X component.
}
\description{
Temporal Variance Component (Time Dimension)
}
