% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{XeeXhC_Lg}
\alias{XeeXhC_Lg}
\title{Calculate Correlation Matrix for Large Model (slower, but avoids dist-matrix in memory)}
\usage{
XeeXhC_Lg(M, cutoff, X, e, n1, k, kernel = "bartlett", dist_fn = "Haversine")
}
\arguments{
\item{M}{matrix of locations}

\item{cutoff}{the distance for cutoff}

\item{X}{model matrix}

\item{e}{vector of residuals}

\item{n1}{number of observations}

\item{k}{number of regressors}

\item{kernel}{(string) kernel function (default is bartlett-triangular)}

\item{dist_fn}{(string) distance function (Haversine)}
}
\description{
Calculate Correlation Matrix for Large Model (slower, but avoids dist-matrix in memory)
}
