% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conley_VCV.R
\name{compute_conley_lfe}
\alias{compute_conley_lfe}
\title{Compute Conley Standard Error}
\usage{
compute_conley_lfe(lfeobj, cutoff, kernel_choice = "bartlett", ...)
}
\arguments{
\item{...}{}
}
\description{
Convenience function which nests \code{conley_SE} and prints only the standard error.
The same parameters can be used here.
}
