% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conley_VCV.R
\name{conley_SE}
\alias{conley_SE}
\title{Conley Spatial Variance-Covariances for Fixed Effects models fit with lfe}
\usage{
conley_SE(
  reg,
  unit,
  time,
  lat,
  lon,
  kernel = "bartlett",
  dist_fn = "Haversine",
  dist_cutoff = 500,
  lag_cutoff = 5,
  lat_scale = 111,
  verbose = FALSE,
  cores = 1,
  balanced_pnl = FALSE
)
}
\arguments{
\item{balanced_pnl}{}
}
\description{
This is the function containing the underlying C++ code by Darin Christensen molten into a package.
}
