% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_as_cols.R
\name{coords_as_columns}
\alias{coords_as_columns}
\title{Coordinates as columns into a data frame}
\usage{
coords_as_columns(x, names = c("x", "y"))
}
\arguments{
\item{names}{}
}
\description{
This is based on the \code{sfc_as_cols} function which has been flying around on Stackoverflow and other websites for years.
Since there is no package implementation so far, it has been brought into this package.
The function \code{sf::st_coordinates()} achieves something very similar.
}
