% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DistMat}
\alias{DistMat}
\title{Create Distance Matrix}
\usage{
DistMat(M, cutoff, kernel = "bartlett", dist_fn = "Haversine")
}
\arguments{
\item{M}{a matrix of locations (n x 2, latitude and longitude)}

\item{cutoff}{the distance cutoff (bandwidth) in km}

\item{kernel}{(string) kernel function (default is bartlett-triangular)}

\item{dist_fn}{(string) distance function (Haversine)}
}
\value{
A symmetric n x n matrix of kernel weights with 1s on the diagonal.
}
\description{
Create Distance Matrix
}
