% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_corr_range.R
\name{extract_corr_range}
\alias{extract_corr_range}
\title{Extract Correlation Range}
\usage{
extract_corr_range(input, returnzeroifNA = FALSE)
}
\arguments{
\item{input}{a correlogram estimated with ncf::correlog() or a covariogram estimated with gstat::variogram(..., covariogram = TRUE)}
}
\value{
an integer that represents the correlation range (in units of input, i.e. with lon lat it is km, with equal area CRS it typically is meters)
}
\description{
Function will pick the point where the correlogram first crosses the x-axis (the distance), i.e. where the spatial autocorrelation gets 0 for the first time.
This is a good approximation of the actual correlation range.
}
