% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity_centroid.R
\name{gravity_centroid}
\alias{gravity_centroid}
\title{Center of Gravity}
\usage{
gravity_centroid(df.sf, weight = NA)
}
\arguments{
\item{weight}{}
}
\value{

}
\description{
Center of Gravity
}
\references{
Arlinghaus
}
