% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conley_VCV.R
\name{iterateObs}
\alias{iterateObs}
\title{Iterate observations}
\usage{
iterateObs(
  dt,
  Xvars,
  d,
  k,
  sub_index,
  type,
  cutoff,
  balanced_pnl,
  verbose,
  kernel,
  dist_fn
)
}
\description{
Iterate observations
}
