% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_sac.R
\name{lm_sac}
\alias{lm_sac}
\title{Linear model taking into account spatial autocorrelation}
\usage{
lm_sac(
  formula.chr,
  data.sf,
  knn_number = 20,
  conley_cutoff = 5,
  conley_kernel = "bartlett",
  correlograms = F,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
ouput is lm.sac/custom object that can be displayed with modelsummary
}
