% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialInference-package.R
\docType{package}
\name{SpatialInference-package}
\alias{SpatialInference}
\alias{SpatialInference-package}
\title{SpatialInference: Tools for Statistical Inference with Geo-Coded Data}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/axlehner/SpatialInference}
  \item Report bugs at \url{https://github.com/axlehner/SpatialInference/issues}
}

}
\author{
\strong{Maintainer}: Alexander Lehner \email{lehner@uchicago.edu} (\href{https://orcid.org/0000-0001-5885-5966}{ORCID})

}
\keyword{internal}
