% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TimeDist}
\alias{TimeDist}
\title{Calculate Correlation Matrix (time-dim)}
\usage{
TimeDist(times, cutoff, X, e, n1, k)
}
\arguments{
\item{times}{vector containing the different times}

\item{cutoff}{the distance for cutoff}

\item{X}{model matrix}

\item{e}{vector of residuals}

\item{n1}{number of observations}

\item{k}{number of regressors}
}
\description{
Calculate Correlation Matrix (time-dim)
}
