% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{US_counties_centroids}
\alias{US_counties_centroids}
\title{Centroids of contiguous US counties (2017)}
\format{
\subsection{\code{US_counties_centroids}}{

A spatial data frame with 3,108 rows:
\describe{
\item{STATE}{Code for the state the county is in}
\item{NAME, NAMELSAD}{Name of county}
\item{GISJOIN}{Unique ID to join with IPUMS data at the county level (2017 geographies)}
...
}
}
}
\source{
IPUMS
}
\usage{
US_counties_centroids
}
\description{
sf data frame with 3108 observations and NAD83 datum
}
\keyword{datasets}
