% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{XeeXhC_Lg}
\alias{XeeXhC_Lg}
\title{Spatial Variance Component for Large Datasets}
\usage{
XeeXhC_Lg(M, cutoff, X, e, n1, k, kernel = "bartlett", dist_fn = "Haversine")
}
\arguments{
\item{M}{matrix of locations (n x 2, latitude and longitude)}

\item{cutoff}{the distance cutoff (bandwidth) in km}

\item{X}{model matrix (n x k)}

\item{e}{vector of residuals (length n)}

\item{n1}{number of observations}

\item{k}{number of regressors}

\item{kernel}{(string) kernel function (default is bartlett-triangular)}

\item{dist_fn}{(string) distance function (Haversine)}
}
\value{
A k x k matrix representing the spatial X'ee'X component.
}
\description{
Memory-efficient variant that avoids constructing the full n x n distance matrix.
}
